// src/models/user-activity-type.ts
export enum UserActivityType {
    Registered = 'Registered',
    RegistrationCompleted = 'RegistrationCompleted',
    EmailVerified = 'EmailVerified',
    AdditionalEmailVerified = 'AdditionalEmailVerified',
    AdditionalEmailDeleted = 'AdditionalEmailDeleted',
    IdVerified = 'IdVerified',
    PaymentStatusChanged = 'PaymentStatusChanged',
    StripeAccountStatusUpdated = 'StripeAccountStatusUpdated',
    LinkedAccountCreated = 'LinkedAccountCreated',
    PasswordRecovered = 'PasswordRecovered',
    ProfileEdited = 'ProfileEdited',
    LinkedAccountProfileEdited = 'LinkedAccountProfileEdited',
    Blocked = 'Blocked',
    UnBlocked = 'UnBlocked',
    LinkedAccountDeleted = 'LinkedAccountDeleted',
    Deleted = 'Deleted'
} 